Replace one element of each row with a random element from given data frame.

Replace one element of each row with a random element from given data frame.

replace_items_by_rows(chosen_vars, data, target)

Arguments

chosen_vars
Columns to sample from [as positions/numbers].

data
Data frame to sample from.

target
Data frame in which the replacement is done.

Value

data.frame