Generate dataset for local exploration. DEPRECATED. Please refer to sample_locally2 function for updated and improved interface. This function will be removed in the future and was left only to remain consistent with the examples given in https://arxiv.org/abs/1804.01955. For more, see NEWS and vignette.

Generate dataset for local exploration. DEPRECATED. Please refer to sample_locally2 function for updated and improved interface. This function will be removed in the future and was left only to remain consistent with the examples given in https://arxiv.org/abs/1804.01955. For more, see NEWS and vignette.

sample_locally(data, explained_instance, explained_var, size,
  standardise = FALSE)

Arguments

data
Data frame from which new dataset will be simulated.

explained_instance
One row data frame with the same variables as in data argument. Local exploration will be performed around this observation.

explained_var
Name of a column with the variable to be predicted.

size
Number of observations is a simulated dataset.

standardise
If TRUE, numerical variables will be scaled to have mean 0, var 1.

Value

list

Examples

# NOT RUN {
dataset_for_local_exploration <- sample_locally(data = wine,
                                               explained_instance = wine[5, ],
                                               explained_var = "quality",
                                               size = 50,
                                               standardise = TRUE)
# }